# labassist-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับระบบจัดการ TA ภาควิชา CS มหาวิทยาลัยศิลปากร",
        "title": "LabAssist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/courses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "รายการวิชาทั้งหมด (สาธารณะ)",
                "parameters": [
                    {
                        "enum": [
                            "open",
                            "closing_soon",
                            "closed",
                            "draft"
                        ],
                        "type": "string",
                        "description": "กรองตามสถานะ",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค้นหาด้วยชื่อหรือรหัสวิชา",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Course"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "ดูรายละเอียดวิชา (สาธารณะ)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Course"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/instructor/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "รายการวิชาของอาจารย์",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Course"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "สร้างวิชาใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลวิชา",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Course"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instructor/courses/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "แก้ไขข้อมูลวิชา",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ข้อมูลที่ต้องการแก้ไข",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Course"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instructor/courses/{id}/applicants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "รายชื่อผู้สมัครของวิชา",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ta",
                            "labboy"
                        ],
                        "type": "string",
                        "description": "กรองตามประเภทที่สมัคร",
                        "name": "role_applied",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "pending",
                            "accepted",
                            "rejected",
                            "withdrawn"
                        ],
                        "type": "string",
                        "description": "กรองตามสถานะ",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค้นหาด้วยชื่อหรือรหัสนักศึกษา",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Application"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instructor/courses/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "อัพเดตสถานะวิชา",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "สถานะใหม่",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateCourseStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Course"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateCourseRequest": {
            "type": "object",
            "required": [
                "academic_year",
                "code",
                "semester",
                "title"
            ],
            "properties": {
                "academic_year": {
                    "type": "integer",
                    "example": 2567
                },
                "code": {
                    "type": "string",
                    "example": "CS101"
                },
                "description": {
                    "type": "string"
                },
                "labboy_slots": {
                    "type": "integer",
                    "example": 2
                },
                "requirements": {
                    "type": "string"
                },
                "semester": {
                    "type": "string",
                    "example": "1"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CourseStatus"
                        }
                    ],
                    "example": "draft"
                },
                "ta_slots": {
                    "type": "integer",
                    "example": 3
                },
                "title": {
                    "type": "string",
                    "example": "Introduction to Programming"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "handlers.UpdateCourseRequest": {
            "type": "object",
            "properties": {
                "academic_year": {
                    "type": "integer",
                    "example": 2567
                },
                "code": {
                    "type": "string",
                    "example": "CS101"
                },
                "description": {
                    "type": "string"
                },
                "labboy_slots": {
                    "type": "integer",
                    "example": 2
                },
                "requirements": {
                    "type": "string"
                },
                "semester": {
                    "type": "string",
                    "example": "1"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CourseStatus"
                        }
                    ],
                    "example": "open"
                },
                "ta_slots": {
                    "type": "integer",
                    "example": 3
                },
                "title": {
                    "type": "string",
                    "example": "Introduction to Programming"
                }
            }
        },
        "handlers.UpdateCourseStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CourseStatus"
                        }
                    ],
                    "example": "open"
                }
            }
        },
...
        "models.AppStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected",
                "withdrawn"
            ],
            "x-enum-varnames": [
                "AppPending",
                "AppAccepted",
                "AppRejected",
                "AppWithdrawn"
            ]
        },
        "models.Application": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "course_code": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "course_title": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "motivation": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "reviewed_at": {
                    "type": "string"
                },
                "reviewed_by_id": {
                    "type": "integer"
                },
                "reviewed_by_name": {
                    "type": "string"
                },
                "role_applied": {
                    "$ref": "#/definitions/models.RoleApplied"
                },
                "status": {
                    "$ref": "#/definitions/models.AppStatus"
                },
                "student_code": {
                    "type": "string"
                },
                "student_email": {
                    "type": "string"
                },
                "student_faculty": {
                    "type": "string"
                },
                "student_gpa": {
                    "type": "number"
                },
                "student_id": {
                    "type": "integer"
                },
                "student_name": {
                    "description": "Computed fields (not in DB)",
                    "type": "string"
                },
                "student_year": {
                    "type": "integer"
                }
            }
        },
        "models.Course": {
            "type": "object",
            "properties": {
                "academic_year": {
                    "type": "integer"
                },
                "applicant_count": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructor_id": {
                    "type": "integer"
                },
                "instructor_name": {
                    "type": "string"
                },
                "labboy_accepted": {
                    "type": "integer"
                },
                "labboy_slots": {
                    "type": "integer"
                },
                "requirements": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.CourseStatus"
                },
                "ta_accepted": {
                    "type": "integer"
                },
                "ta_slots": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CourseStatus": {
            "type": "string",
            "enum": [
                "open",
                "closing_soon",
                "closed",
                "draft"
            ],
            "x-enum-varnames": [
                "StatusOpen",
                "StatusClosingSoon",
                "StatusClosed",
                "StatusDraft"
            ]
        },
        "models.RoleApplied": {
            "type": "string",
            "enum": [
                "ta",
                "labboy"
            ],
            "x-enum-varnames": [
                "RoleTA",
                "RoleLabBoy"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}